Moved more params to analyses data entry point
