+++ lib/assets/javascripts/cartodb3/components/modals/add-analysis/analysis-info-pane.js
@@ -1,6 +1,6 @@
 var CoreView = require('backbone/core-view');
 var template = require('./analysis-info-pane.tpl');
-var ANALYSIS_ANIMATION_TEMPLATES = require('./analysis-animation-templates');
+var Analyses = require('../../../data/analyses');
 
 /**
  * View to render an individual analysis category and its analysis options
@@ -37,11 +37,12 @@ module.exports = CoreView.extend({
     if (selectedAnalysis) {
       var analysisParams = {};
       analysisType = selectedAnalysis.get('type');
-      var animationTemplate = ANALYSIS_ANIMATION_TEMPLATES[analysisType];
+      var analysisItem = Analyses.getAnalysisByType(analysisType);
+      var animationTemplate = analysisItem && analysisItem.animationTemplate;
 
       if (animationTemplate) {
-        genericType = animationTemplate.genericType || analysisType;
-        analysisParams = animationTemplate.params || {};
+        genericType = analysisItem.genericType || analysisType;
+        analysisParams = analysisItem.animationParams || {};
       }
 
       this._analysisCategoryName = selectedAnalysis.get('category');
