+++ lib/assets/javascripts/cartodb3/components/modals/add-analysis/analysis-option-view.js
@@ -1,7 +1,7 @@
 var _ = require('underscore');
 var CoreView = require('backbone/core-view');
 var template = require('./analysis-option.tpl');
-var ANALYSIS_ANIMATION_TEMPLATES = require('./analysis-animation-templates');
+var Analyses = require('../../../data/analyses');
 
 /**
  * View for an individual analysis option.
@@ -33,17 +33,18 @@ module.exports = CoreView.extend({
     }
 
     var analysisType = this.model.get('type');
-    var animationTemplate = ANALYSIS_ANIMATION_TEMPLATES[analysisType];
+    var analysisItem = Analyses.getAnalysisByType(analysisType);
+    var animationTemplate = analysisItem && analysisItem.animationTemplate;
     var genericType = '';
 
     if (animationTemplate) {
-      genericType = animationTemplate.genericType || analysisType;
+      genericType = analysisItem.genericType || analysisType;
     }
 
     this.$el.html(template(_.extend(props, { type: genericType })));
 
     if (animationTemplate) {
-      this.$('.js-animation').append(animationTemplate.template);
+      this.$('.js-animation').append(animationTemplate);
     }
 
     this.$el.toggleClass('is-selected', this.model.get('selected'));
