+++ lib/assets/javascripts/cartodb3/components/onboardings/analysis/analysis-launcher.js
@@ -1,9 +1,7 @@
 var LocalStorage = require('../../../components/local-storage/local-storage');
 var AnalysisOnboardingView = require('../../../components/onboardings/analysis/analysis-view');
-
+var Analyses = require('../../../data/analyses');
 var STORAGE_KEY = 'onboarding';
-
-var ANALYSES_TYPES = require('./analyses-types');
 
 var launcher = (function () {
   var onboardings;
@@ -40,7 +38,7 @@ var launcher = (function () {
     },
 
     launch: function (type, model) {
-      var genericType = ANALYSES_TYPES[type] && ANALYSES_TYPES[type].genericType || type;
+      var genericType = Analyses.getAnalysisByType(type).genericType || type;
 
       if (!LocalStorage.get(genericType) && genericType) {
         onboardings.create(function (modalModel) {
