+++ lib/assets/javascripts/cartodb3/components/onboardings/analysis/analysis-view.js
@@ -1,9 +1,9 @@
 var LocalStorage = require('../../../components/local-storage/local-storage');
 var CoreView = require('backbone/core-view');
 var template = require('./analysis-view.tpl');
+var Analyses = require('../../../data/analyses');
 
 var STORAGE_KEY = 'onboarding';
-var ANALYSES_TYPES = require('./analyses-types');
 
 module.exports = CoreView.extend({
   className: 'AnalysisCompletionDetails is-opening',
@@ -34,7 +34,8 @@ module.exports = CoreView.extend({
       type: this._getGenericAnalysisType()
     }));
 
-    var html = this._typeDef().template(this.model.attributes);
+    var onboardingTemplate = this._typeDef().onboardingTemplate;
+    var html = onboardingTemplate && onboardingTemplate(this.model.attributes);
     this.$('.js-content').html(html);
 
     this._onChangeEditorModelEdition(this._editorModel);
@@ -75,11 +76,7 @@ module.exports = CoreView.extend({
 
   _typeDef: function (type) {
     type = type || this.model.get('type');
-    return ANALYSES_TYPES[type];
-  },
-
-  _getStyleAnalysisAfterFinish: function () {
-    return this._typeDef().style;
+    return Analyses.getAnalysisByType(type);
   },
 
   _getGenericAnalysisType: function () {
