+++ lib/assets/javascripts/cartodb3/data/analyses.js
@@ -15,40 +15,60 @@ var MAP = {
     FormModel: AggregateIntersectionFormModel,
     modalTitle: _t('components.modals.add-analysis.option-types.aggregate-intersection.title'),
     modalDesc: _t('components.modals.add-analysis.option-types.aggregate-intersection.desc'),
-    modalCategory: 'data_transformation'
+    modalCategory: 'data_transformation',
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/aggregate-intersection.tpl'),
+    animationTemplate: require('../components/modals/add-analysis/animation-templates/intersect.tpl')
   },
   'bounding-box': {
     title: _t('analyses.bounding-box.short-title'),
-    FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/group-points-form-model')
+    FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/group-points-form-model'),
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/group-points.tpl'),
+    genericType: 'group-points'
   },
   'bounding-circle': {
     title: _t('analyses.bounding-circle.short-title'),
-    FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/group-points-form-model')
+    FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/group-points-form-model'),
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/group-points.tpl'),
+    genericType: 'group-points'
   },
   'buffer': {
     title: _t('analyses.area-of-influence.short-title'),
     FormModel: AreaOfInfluenceFormModel,
     modalTitle: _t('components.modals.add-analysis.option-types.area-of-influence.title'),
     modalDesc: _t('components.modals.add-analysis.option-types.area-of-influence.desc'),
-    modalCategory: 'data_transformation'
+    modalCategory: 'data_transformation',
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/area-of-influence.tpl'),
+    animationTemplate: require('../components/modals/add-analysis/animation-templates/aoi.tpl'),
+    genericType: 'area-of-influence'
   },
   'centroid': {
     title: _t('analyses.centroid.short-title'),
     FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/centroid-form-model'),
     modalTitle: _t('components.modals.add-analysis.option-types.centroid.title'),
     modalDesc: _t('components.modals.add-analysis.option-types.centroid.desc'),
-    modalCategory: 'data_transformation'
+    modalCategory: 'data_transformation',
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/centroid.tpl'),
+    animationTemplate: require('../components/modals/add-analysis/animation-templates/centroid.tpl'),
+    genericType: 'centroid'
   },
   'convex-hull': {
     title: _t('analyses.convex-hull.short-title'),
     FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/group-points-form-model'),
     modalTitle: _t('components.modals.add-analysis.option-types.group-points.title'),
     modalDesc: _t('components.modals.add-analysis.option-types.group-points.desc'),
-    modalCategory: 'data_transformation'
+    modalCategory: 'data_transformation',
+    onboardingTemplate: require('../components/onboardings/analysis/analyses//group-points.tpl'),
+    animationTemplate: require('../components/modals/add-analysis/animation-templates/group-points.tpl'),
+    animationParams: {
+      method: _t('analyses-onboarding.placeholders.method')
+    },
+    genericType: 'group-points'
   },
   'concave-hull': {
     title: _t('analyses.concave-hull.short-title'),
-    FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/group-points-form-model')
+    FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/group-points-form-model'),
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/group-points.tpl'),
+    genericType: 'group-points'
   },
   'data-observatory-measure': {
     title: _t('analyses.data-observatory-measure.short-title'),
@@ -58,46 +78,65 @@ var MAP = {
     modalCategory: 'create_clean',
     checkIfValid: function (userModel, configModel) {
       return checkIfValidByCustomInstall(userModel, configModel);
-    }
+    },
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/data-observatory-measure.tpl'),
+    animationTemplate: require('../components/modals/add-analysis/animation-templates/data-observatory-measure.tpl'),
+    genericType: 'data-observatory-measure'
   },
   'filter-by-node-column': {
     title: _t('analyses.filter-by-node-column.short-title'),
     FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/filter-by-node-column'),
     modalTitle: _t('components.modals.add-analysis.option-types.filter-by-node-column.title'),
     modalDesc: _t('components.modals.add-analysis.option-types.filter-by-node-column.desc'),
-    modalCategory: 'create_clean'
+    modalCategory: 'create_clean',
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/filter-by-node-column.tpl'),
+    animationTemplate: require('../components/modals/add-analysis/animation-templates/filter-by-layer.tpl'),
+    animationParams: {
+      name_of_the_layer: _t('analyses-onboarding.placeholders.layer-name')
+    }
   },
   'filter-category': {
     title: _t('analyses.filter.short-title'),
-    FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/filter-form-model')
+    FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/filter-form-model'),
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/filter.tpl'),
+    genericType: 'filter'
   },
   'filter-range': {
     title: _t('analyses.filter.short-title'),
     FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/filter-form-model'),
     modalTitle: _t('components.modals.add-analysis.option-types.filter.title'),
     modalDesc: _t('components.modals.add-analysis.option-types.filter.desc'),
-    modalCategory: 'data_transformation'
+    modalCategory: 'data_transformation',
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/filter.tpl'),
+    animationTemplate: require('../components/modals/add-analysis/animation-templates/filter-by-column-value.tpl'),
+    genericType: 'filter'
   },
   'georeference-city': {
     title: _t('analyses.georeference.short-title'),
     FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/georeference-form-model'),
     checkIfValid: function (userModel, configModel) {
       return checkIfValidByCustomInstall(userModel, configModel);
-    }
+    },
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/georeference.tpl'),
+    genericType: 'georeference'
   },
   'georeference-ip-address': {
     title: _t('analyses.georeference.short-title'),
     FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/georeference-form-model'),
     checkIfValid: function (userModel, configModel) {
       return checkIfValidByCustomInstall(userModel, configModel);
-    }
+    },
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/georeference.tpl'),
+    genericType: 'georeference'
   },
   'georeference-country': {
     title: _t('analyses.georeference.short-title'),
     FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/georeference-form-model'),
     checkIfValid: function (userModel, configModel) {
       return checkIfValidByCustomInstall(userModel, configModel);
-    }
+    },
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/georeference.tpl'),
+    genericType: 'georeference'
   },
   'georeference-long-lat': {
     title: _t('analyses.georeference.short-title'),
@@ -105,6 +144,8 @@ var MAP = {
     modalTitle: _t('components.modals.add-analysis.option-types.georeference.title'),
     modalDesc: _t('components.modals.add-analysis.option-types.georeference.desc'),
     modalCategory: 'create_clean',
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/georeference.tpl'),
+    animationTemplate: require('../components/modals/add-analysis/animation-templates/georeference.tpl'),
     genericType: 'georeference'
   },
   'georeference-postal-code': {
@@ -112,7 +153,9 @@ var MAP = {
     FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/georeference-form-model'),
     checkIfValid: function (userModel, configModel) {
       return checkIfValidByCustomInstall(userModel, configModel);
-    }
+    },
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/georeference.tpl'),
+    genericType: 'georeference'
   },
   'georeference-street-address': {
     title: _t('analyses.georeference.short-title'),
@@ -120,47 +163,70 @@ var MAP = {
     checkIfValid: function (userModel, configModel) {
       if (!userModel) throw new Error('userModel is required');
       return userModel.get('geocoder_provider');
-    }
+    },
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/georeference.tpl'),
+    genericType: 'georeference'
   },
   'georeference-admin-region': {
     title: _t('analyses.georeference.short-title'),
     FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/georeference-form-model'),
     checkIfValid: function (userModel, configModel) {
       return checkIfValidByCustomInstall(userModel, configModel);
-    }
+    },
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/georeference.tpl'),
+    genericType: 'georeference'
   },
   'intersection': {
     title: _t('analyses.filter-intersection.short-title'),
     FormModel: FilterIntersectionFormModel,
     modalTitle: _t('components.modals.add-analysis.option-types.filter-intersection.title'),
     modalDesc: _t('components.modals.add-analysis.option-types.filter-intersection.desc'),
-    modalCategory: 'data_transformation'
+    modalCategory: 'data_transformation',
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/intersection.tpl'),
+    animationTemplate: require('../components/modals/add-analysis/animation-templates/filter-by-polygon.tpl')
   },
   'kmeans': {
     title: _t('analyses.kmeans.short-title'),
     FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/kmeans-form-model'),
     modalTitle: _t('components.modals.add-analysis.option-types.kmeans.title'),
     modalDesc: _t('components.modals.add-analysis.option-types.kmeans.desc'),
-    modalCategory: 'analyze_predict'
+    modalCategory: 'analyze_predict',
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/kmeans.tpl'),
+    animationTemplate: require('../components/modals/add-analysis/animation-templates/kmeans.tpl'),
+    animationParams: {
+      method: _t('analyses-onboarding.placeholders.clusters')
+    }
   },
   'line-to-column': {
     title: _t('analyses.connect-with-lines.short-title'),
-    FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/connect-with-lines-form-model')
+    FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/connect-with-lines-form-model'),
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/connect-with-lines.tpl'),
+    animationTemplate: require('../components/modals/add-analysis/animation-templates/connect-with-lines.tpl'),
+    genericType: 'connect-with-lines'
   },
   'line-to-single-point': {
     title: _t('analyses.connect-with-lines.short-title'),
     FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/connect-with-lines-form-model'),
     modalTitle: _t('components.modals.add-analysis.option-types.connect-with-lines.title'),
     modalDesc: _t('components.modals.add-analysis.option-types.connect-with-lines.desc'),
-    modalCategory: 'data_transformation'
+    modalCategory: 'data_transformation',
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/connect-with-lines.tpl'),
+    animationTemplate: require('../components/modals/add-analysis/animation-templates/connect-with-lines.tpl'),
+    genericType: 'connect-with-lines'
   },
   'line-source-to-target': {
     title: _t('analyses.connect-with-lines.short-title'),
-    FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/connect-with-lines-form-model')
+    FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/connect-with-lines-form-model'),
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/connect-with-lines.tpl'),
+    animationTemplate: require('../components/modals/add-analysis/animation-templates/connect-with-lines.tpl'),
+    genericType: 'connect-with-lines'
   },
   'line-sequential': {
     title: _t('analyses.connect-with-lines.short-title'),
-    FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/connect-with-lines-form-model')
+    FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/connect-with-lines-form-model'),
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/connect-with-lines.tpl'),
+    animationTemplate: require('../components/modals/add-analysis/animation-templates/connect-with-lines.tpl'),
+    genericType: 'connect-with-lines'
   },
   'merge': {
     title: _t('analyses.merge.short-title'),
@@ -168,14 +234,18 @@ var MAP = {
     modalTitle: _t('components.modals.add-analysis.option-types.merge.title'),
     modalDesc: _t('components.modals.add-analysis.option-types.merge.desc'),
     ModalModel: MergeAnalysisOptionModel,
-    modalCategory: 'create_clean'
+    modalCategory: 'create_clean',
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/merge.tpl'),
+    animationTemplate: require('../components/modals/add-analysis/animation-templates/join-two-columns.tpl')
   },
   'moran': {
     title: _t('analyses.moran-cluster.short-title'),
     FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/moran-form-model'),
     modalTitle: _t('components.modals.add-analysis.option-types.moran-cluster.title'),
     modalDesc: _t('components.modals.add-analysis.option-types.moran-cluster.desc'),
-    modalCategory: 'analyze_predict'
+    modalCategory: 'analyze_predict',
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/moran.tpl'),
+    animationTemplate: require('../components/modals/add-analysis/animation-templates/outliers.tpl')
   },
   'routing-sequential': {
     checkIfValid: function (userModel, configModel) {
@@ -197,14 +267,22 @@ var MAP = {
     FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/sampling-form-model'),
     modalTitle: _t('components.modals.add-analysis.option-types.sampling.title'),
     modalDesc: _t('components.modals.add-analysis.option-types.sampling.desc'),
-    modalCategory: 'data_transformation'
+    modalCategory: 'data_transformation',
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/sampling.tpl'),
+    animationTemplate: require('../components/modals/add-analysis/animation-templates/sampling.tpl'),
+    animationParams: {
+      percentage: _t('analyses-onboarding.placeholders.percentage')
+    },
+    genericType: 'sampling'
   },
   'spatial-markov-trend': {
     title: _t('analyses.spatial-markov-trend.short-title'),
     FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/spatial-markov-trend-model'),
     modalTitle: _t('components.modals.add-analysis.option-types.spatial-markov-trend.title'),
     modalDesc: _t('components.modals.add-analysis.option-types.spatial-markov-trend.desc'),
-    modalCategory: 'analyze_predict'
+    modalCategory: 'analyze_predict',
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/spatial-markov-trend.tpl'),
+    animationTemplate: require('../components/modals/add-analysis/animation-templates/predict-trends.tpl')
   },
   'trade-area': {
     title: _t('analyses.area-of-influence.short-title'),
@@ -212,11 +290,15 @@ var MAP = {
     checkIfValid: function (userModel, configModel) {
       if (!userModel) throw new Error('userModel is required');
       return !!userModel.get('isolines_provider');
-    }
+    },
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/area-of-influence.tpl'),
+    genericType: 'area-of-influence'
   },
   'weighted-centroid': {
     title: _t('analyses.centroid.short-title'),
-    FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/centroid-form-model')
+    FormModel: require('../editor/layers/layer-content-views/analyses/analysis-form-models/centroid-form-model'),
+    onboardingTemplate: require('../components/onboardings/analysis/analyses/centroid.tpl'),
+    genericType: 'centroid'
   }
 };
